jobs-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.4.2
apscheduler==3.10.4
httpx==0.25.2
pandas==2.3.2
numpy>=2.2.6
pandas-ta==0.4.71b0